gsdx: use size_t for loop index when it used countof macro * fix override_GL_ARB_copy_image typo
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@5687 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/GSdx/GPUState.cpp
@@ -27,7 +27,7 @@ GPUState::GPUState()
 {
 	memset(m_status, 0, sizeof(m_status));
 
-	for(uint32 i = 0; i < countof(m_fpGPUStatusCommandHandlers); i++)
+	for(size_t i = 0; i < countof(m_fpGPUStatusCommandHandlers); i++)
 	{
 		m_fpGPUStatusCommandHandlers[i] = &GPUState::SCH_Null;
 	}
